# backend/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a sample server api for beacon based attendance system"
  version: "1.0.0"
  title: "Swagger beacon-based attendance system"
  contact:
    email: "mdsazzadul.islam@tuni.fi"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "127.0.0.1:8080"
basePath: "/v1"
paths:
...
  /event:
    get:
      tags:
      - "Event"
      summary: "Get all the events"
      description: "Get all the events"
      operationId: "getAllEvents"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: string
      responses:
        200:
          description: Get all events
          schema:
            $ref: '#/definitions/userEventList'
        401:
          description: Authorization failed
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
    post:
      tags:
        - "Event"
      summary: "add a new event"
      description: "Add a new event"
      operationId: "addEvent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Adding a new event"
        required: true
        schema:
          $ref: "#/definitions/newEvent"  
      responses:
        201:
          description: event created
        400:
          description : Invalid data
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
  /event/eventStatus/{eventId}:
    put:
      tags:
      - "Event"
      summary: Active/deactive the event
      description: "Teacher can active/deactive the event. When the event is active, student can give attendance."
      operationId: "changeEventStatus"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "path"
        name: "eventId"
        required: true
        type: "integer"
        minimum: 1
        format: "int32"
        description: "ID of the event"
      responses:
        201:
          description: event status changed
          schema:
            type: "object"
            properties:
              status:
                type: boolean
              message:
                type: string
        400:
          description : Invalid data
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
  /event/getEventsByEventPersonal:
    get:
      tags:
        - "Event"
      description: "Get all the event assigned to the user"
      summary: "Show all the events assigned by the user"
      operationId: "getAllEventsAssignedToUser"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      responses:
        200:
          description: Ok
          schema: 
            $ref: "#/definitions/userEventList"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
  /event/{eventId}:
    get:
      tags: 
      - "Event"
      summary: "Show the detail of an individual event"
      description: "Get all the event" 
      operationId: "getEventbyId"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - name: "eventId"
        in: "path"
        required: true
        type: "integer"
        minimum: 1
        format: "int64"
        description: "ID of the event"
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/eventDetails'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
    put:
      tags: 
        - "Event"
      summary: "Update and event"
      description: "Update and event"
      operationId: "updateEvent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "path"
        name: "eventId"
        required: true
        type: "integer"
        minimum: 1
        format: "int32"
        description: "ID of the event"
      - in: "body"
        name: "body"
        description: "update an event details by the ID"
        required: true
        schema:
          $ref: "#/definitions/updateEventBody"
      responses:
        201:
          description: event updated
        400:
          description : Invalid data
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
    delete:
      tags: 
        - "Event"
      summary: "delete an event"
      description: Delete an event. Either person who created the event or who is assigned to take the event can delete.
      operationId: "deleteEvent"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - name: "eventId"
        in: "path"
        required: true
        type: "integer"
        minimum: 1
        format: "int64"
        description: "ID of the event"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid data"
          schema:
            $ref: '#/definitions/responseWithMessage'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
  /event/attendanceResult:
    get:
      tags:
      - Event
      summary: "Get student list from a event attendance"
      description: "Get student list from a event attendance" 
      operationId: "attendanceResult"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - name: "eventId"
        in: "query"
        type: "integer"
        minimum: 1
        format: "int64"
        description: "ID of the event"
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/eventAttendanceStudentInfo'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
  /event/course/{courseId}:
    get:
      tags:
      - Event
      summary: "Get events list for a course"
      description: "Get events list for a course"
      operationId: "courseEvents"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - name: "courseId"
        in: "path"
        type: "integer"
        required: true
        minimum: 1
        format: "int64"
        description: "ID of the course"
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/userEventList'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
  /eventSumByPersonal:
    get:
      tags:
      - "Event"
      summary: returns events summary assigned for a course personal
      description: Returns events summary assigned for a course personal
      operationId: eventSumByPersonal
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/eventSumByPersonal'
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      security:
        - api_key: []
      x-swagger-router-controller: "Event"
...
definitions:
  '404':
    type: object
    additionalProperties: false
    properties:
      error:
        type: string
    required:
      - error
  responseWithMessage:
    type: object
    additionalProperties: false
    properties:
      message:
        type: string
    required:
      - message
...
  newEvent:
    type: object
    properties:
      courseId:
        type: integer
      eventPersonal:
        type: integer  
      eventName:
        type: string
        minLength: 5
      eventType:
        type: string
        enum: ['class', 'exercise']
      roomId:
        type: integer
      startDate:
        description: Start date of event
        type: string
        pattern: '(^\d{4}-(?:0[0-9]{1}|1[0-2]{1})-[0-9]{2})$'
      endDate:
        description: End date of event
        type: string
        pattern: '(^\d{4}-(?:0[0-9]{1}|1[0-2]{1})-[0-9]{2})$'
      startTime:
        description: Start time of event
        type: string
        pattern: '(^\d{2}:\d{2}:\d{2})$'
      endTime:
        description: End time of event
        type: string
        pattern: '(^\d{2}:\d{2}:\d{2})$'
      recurrent:
        description: recurrent events
        type: boolean
        default: false
      everyAfter:
        description: repeat event every after given number
        type: number
      everyAfterType:
        description: type of every after
        type: string
        enum: ['day', 'week']
    example:
      courseId: 0
      eventPersonal: 0
      eventName: "DIP Lecture 1"
      eventType: "class"
      roomId: 0
      startDate: "2019-11-10"
      endDate: "2019-11-10"
      startTime: "14:00:19"
      endTime: "16:00:19"
      recurrent: false
      everyAfter: 1
      everyAfterType: 'day'
    additionalProperties: false
    required:
      - courseId
      - eventPersonal
      - eventName
      - eventType
      - roomId
      - startDate
      - endDate
      - startTime
      - endTime
  userEventList:
    type: "array"
    items:
      $ref: "#/definitions/eventDetails"
  eventDetails:
    type: object
    properties:
      id:
        type: integer
      courseId:
        type: integer
      courseCode:
        type: string
      courseName:
        type: string
      eventPersonal:
        type: integer
      firstName:
        type: string
      lastName:
        type: string
      eventname:
        type: string
      eventType:
        type: string
      roomId:
        type: integer
      roomName:
        type: string
      startTime:
        format: date-time
        description: startTime
        type: string
      endTime:
        format: date-time
        description: endtTime
        type: string
      activeStatus:
        type: boolean
    example:
      id: 1
      courseId: 1
      courseCode: 'TIE-00000'
      courseName: 'TEST COURSE'
      eventPersonal: 1
      firstName: 'First name'
      lastName: 'Last name'
      eventname: 'DIP Lecture 1'
      eventType: 'class'
      roomId: 1
      roomName: 'TC-111'
      startDateTime: "2019-11-22 08:00:19"
      endDateTime: "2019-11-22 10:00:19"
    additionalProperties: false
  updateEventBody:
    type: object
    additionalProperties: false
    properties:
      courseId:
        type: integer
      eventPersonal:
        type: integer  
      eventName:
        type: string
      eventType:
        type: string
      roomId:
        type: integer
      startDate:
        description: Start date of event
        type: string
        pattern: '(^\d{4}-(?:0[0-9]{1}|1[0-2]{1})-[0-9]{2})$'
      endDate:
        description: End date of event
        type: string
        pattern: '(^\d{4}-(?:0[0-9]{1}|1[0-2]{1})-[0-9]{2})$'
      startTime:
        description: Start time of event
        type: string
        pattern: '(^\d{2}:\d{2}:\d{2})$'
      endTime:
        description: End time of event
        type: string
        pattern: '(^\d{2}:\d{2}:\d{2})$'
    example:
      courseId: 1
      eventPersonal: 2
      eventName: "DIP Lecture 1"
      eventType: "class"
      roomId: 1
      startDate: "2019-11-10"
      startTime: "14:00:19"
      endDate: "2019-11-10"
      endTime: "15:00:19"
    required:
      - courseId
      - eventPersonal
      - eventName
      - eventType
      - roomId
      - startDate
      - endDate
      - startTime
      - endTime
...
  eventAttendanceStudentInfo:
    type: array
    items:
      type: object
      properties:
        student_id:
          type: integer
          format: int32
        student_name:
          type: string
        attendance_time:
          type: string
        event_name:
          type: string
        event_id:
          type: integer
          format: int32
      required:
      - student_id
      - student_name
      - attendance_time
      - event_name
      - event_id
      additionalProperties: false
    example: [
      {
        student_id: 1,
        student_name: 'Student One',
        attendance_time: '2019-11-22 09:40:19',
        event_name: 'DIP Lecture 1',
        event_id: 3
      },
      {
        student_id: 2,
        student_name: 'Student Two',
        attendance_time: '2019-11-22 09:40:49',
        event_name: 'DIP Lecture 1',
        event_id: 3
      }
    ]
  eventSumByPersonal:
    type: array
    items:
      type: object
      properties:
        course_id:
          type: integer
          format: int32
        course_code:
          type: string
        course_name:
          type: string
        event_type:
          type: array
          items:
            type: object
            properties:
              type_name:
                type: string
              total_event:
                type: integer
                format: int32
              next_event:
                type: string
              room_name:
                type: string
              next_event_id:
                type: integer
                format: int32
      additionalProperties: false
    example:
      [
        {
          "course_id": 21,
          "course_name": "Data Structures and Algorithms",
          "course_code": "TIE-20106",
          "event_type": [
            {
              "type": "class",
              "total_event": 14,
              "next_event": "2020-01-21 14:00:19",
              "room_name": "TC-110",
              "next_event_id": 10
            },
            {
              "type": "exercise",
              "total_event": 14,
              "next_event": "2020-01-21 17:00:00",
              "room_name": "TC-110",
              "next_event_id": 20
            }
          ]
        },
        {
          "course_id": 23,
          "course_name": "WW2",
          "course_code": "TIE-23157",
          "event_type": [
            {
              "type": "class",
              "total_event": 14,
              "next_event": "",
              "room_name": "",
              "next_event_id": ""
            }
          ]
        }
      ]
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "X-API-KEY"
    in: "header"